checker/requirements.txt
enochecker3==0.8.1
uvicorn
gunicorn
faker
httpx
ffmpeg-python
bs4